[feat/be-vote2] 공감 기능 구현

+++ Server/src/main/java/com/e1i5/mentaltal/board/dto/BoardResponseDto.java
@@ -17,8 +17,8 @@ public class BoardResponseDto {
     private String title;
     private String content;
 //    private String tags;
-    private int viewCount;  // 조회수
-    private int voteCount;  // 공감수
+    private long viewCount;  // 조회수
+    private long voteCount;  // 공감수
     private long commentCount;
 
     private String nickName;
